+++ Source/Core/UICommon/CommandLineParse.cpp
@@ -30,7 +30,7 @@ public:
 
     if (audio_backend.size())
       m_values.emplace_back(
-          std::make_tuple(Config::MAIN_DSP_HLE.location, StringFromBool(audio_backend == "HLE")));
+          std::make_tuple(Config::MAIN_DSP_HLE.location, ValueToString(audio_backend == "HLE")));
 
     // Arguments are in the format of <System>.<Section>.<Key>=Value
     for (const auto& arg : args)
